s4 param: change messaging directory name from messaging to msg
This saves 6 chars and as in some installation users tends to install
samba in directory with very long name they have a problem with the
sockets.
It will also ease the pain on buildfarms as we tend to have very long
environnement name (ie. vampire_dc) in path that are already quite long
leading also to errors on unix socket manipulation.

Autobuild-User: Matthieu Patou <mat@samba.org>
Autobuild-Date: Fri Oct 15 08:15:56 UTC 2010 on sn-devel-104