Fix bug #7700 - Improvement of return code of smbclient
Based on an initial patch from H Hasegawa <hasegawa.hiroyuki@fujixerox.co.jp>.
Convert cli_list and associated functions to take calls that return NTSTATUS.

Note that this is different from the code that went into master
as many of the changes that were made to cli_list in master have
not been back-ported to v3-6-test.

Jeremy.