s4-devel: a script to test the Samba4 HOWTO
This provides a script that allows testing of most of the steps of the
Samba4 HOWTO. The big difference between this and 'make test' is that
it test against windows, using pexpect to control windows boxes via
telnet.

The info about VMs and other parameters are in separate conf
files. I've included a sample config file that I use on my laptop.