wintest Evolve wintest to handle it's own BIND nameserver
The changes here start our own copy of BIND, listening on a new
interface that we create at the start of the script.  The user must
specify the IP address etc that this interface will have, and BIND and
Samba will bind to that interface only.

This means that we won't conflict with an existing BIND on the system,
so long as it isn't configured to listen on all interfaces.

We also auto-create the /etc/resolv.conf and restore it at the end of
the script, using the existing server value from the script as a
forwarder.

Andrew Bartlett

Autobuild-User: Andrew Bartlett <abartlet@samba.org>
Autobuild-Date: Tue Nov 23 08:30:34 CET 2010 on sn-devel-104