s3-build: Remove broken RPC modules support.
With the current module support for rpc we're not able to call the init
functions with the rpc callback structure. So init functions and
shutdown functions aren't called. These init functions are needed to
setup pre requirements like migrating the printer databases and register
at the endpoint mapper. The shutdown functions cleanup memory and
deregister from the endpoint mapper.